requirements.txt
antlr4-python3-runtime==4.8
attrs==21.2.0
click==8.0.1
filelock==3.0.12
gitdb==4.0.7
GitPython==3.1.18
grpcio==1.40.0
hydra-core==1.1.1
msgpack==1.0.2
numpy==1.21.2
omegaconf==2.1.1
pkg_resources==0.0.0
protobuf==3.17.3
PyYAML==5.4.1
ray==1.6.0
redis==3.5.3
six==1.16.0
smmap==4.0.0
tqdm==4.62.2
